+++ apm-sniffer/apm-sdk-plugin/connection-pool-commons/src/main/java/org/apache/skywalking/apm/plugin/connectionpool/ConnectionPoolInfo.java
@@ -13,22 +13,13 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
-package org.apache.skywalking.apm.testcase.hikaricp;
+package org.apache.skywalking.apm.plugin.connectionpool;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+public interface ConnectionPoolInfo {
 
-@SpringBootApplication
-public class Application {
+    void recordGetConnectionTime(long time);
 
-    public static void main(String[] args) {
-        try {
-            SpringApplication.run(Application.class, args);
-        } catch (Exception e) {
-            // Never do this
-        }
-    }
+    void recordGetConnectionStatue(boolean failed);
 }