+++ apm-sniffer/apm-sdk-plugin/connection-pool-commons/src/main/java/org/apache/skywalking/apm/plugin/connectionpool/ObjectHadBeenRecycledException.java
@@ -13,22 +13,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
-package org.apache.skywalking.apm.testcase.hikaricp;
+package org.apache.skywalking.apm.plugin.connectionpool;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        try {
-            SpringApplication.run(Application.class, args);
-        } catch (Exception e) {
-            // Never do this
-        }
-    }
+public class ObjectHadBeenRecycledException extends Exception {
 }